add support for matlab 2019a and vivado 2019.1
Supports rfsoc parts in 2019.1 where those parts are supported in that
corresponding version of vivado

updates device tree generation content to build for 2019.1 sdk, so reverting
references from xsa to hdf and changing the device tree param keys to that
version of the repo. Users will need to clone device-tree-xlnx and checkout back
to that version.